data files to read and create JSON object
the main.csv file contains lines with the main data + reference to the other file with some additional details with 261 lines of data. the path to the file need to be changed since it is located somewhere in the file system.
tbl_log_res....csv files are files referred by main.csv file for detailed information. Each line of main file have a reference to one of this file, and i need to create as many JSON objects as lines in the main.csv. So i'm reading this file line by line and getting the data from the main.csv, retrieving the file_name from it and passing to another flow, reading full data from the csv and after that i need to merge the line from the main.csv and the JSON object read from the tbl_log_res....csv file into single JSON object as in result.json